Replace site_size macro constants with constexpr size_t variables, and fix the resulting unsigned/int comparison warnings.  This promotes the use of these (now) variables to 64-bit fixing a variety of 32-bit overflow bugs in the process.  With these changes we can run clover at 72^4 local volume with the host reference tests passing

+++ tests/contract_test.cpp
@@ -120,12 +120,12 @@ int test(int contractionType, QudaPrecision test_prec)
   void *d_result = safe_malloc(2 * V * 16 * data_size);
 
   if (test_prec == QUDA_SINGLE_PRECISION) {
-    for (int i = 0; i < V * spinor_site_size; i++) {
+    for (size_t i = 0; i < V * spinor_site_size; i++) {
       ((float *)spinorX)[i] = rand() / (float)RAND_MAX;
       ((float *)spinorY)[i] = rand() / (float)RAND_MAX;
     }
   } else {
-    for (int i = 0; i < V * spinor_site_size; i++) {
+    for (size_t i = 0; i < V * spinor_site_size; i++) {
       ((double *)spinorX)[i] = rand() / (double)RAND_MAX;
       ((double *)spinorY)[i] = rand() / (double)RAND_MAX;
     }
